Moved metrics to another function and check variables once.
